Fix some QBE codegen + various changes

+++ src/file_util.c
@@ -8,7 +8,7 @@ char* read_file_to_string(char* path) {
     if (file == NULL) return 0;
 
     fseek(file, 0, SEEK_END);
-    long int file_size = ftell(file);
+    size_t file_size = ftell(file);
     fseek(file, 0, SEEK_SET);
 
     char* buffer = malloc(file_size + 1);
